Experimental coding convention.
Kernel-related functions and variables now belong in a table. Should the need for any additional facilities(graphics, filesystem, I/O, etc.) arises, a new table will be added into the kernel table.

Not everything is moved onto this version, of course. I plan to rewrite most if not all of the kernel(again).

Also, can someone please tell me if this is OOP.